+++ test/aqua/test_svm_classical.py
@@ -2,7 +2,7 @@
 
 # This code is part of Qiskit.
 #
-# (C) Copyright IBM 2018, 2019.
+# (C) Copyright IBM 2018, 2020.
 #
 # This code is licensed under the Apache License, Version 2.0. You may
 # obtain a copy of this license in the LICENSE.txt file in the root directory
@@ -14,7 +14,7 @@
 
 """ Test SVM Classical """
 
-from test.aqua.common import QiskitAquaTestCase
+from test.aqua import QiskitAquaTestCase
 import numpy as np
 from qiskit.aqua import aqua_globals
 from qiskit.aqua.algorithms import SVM_Classical
@@ -98,12 +98,14 @@ class TestSVMClassical(QiskitAquaTestCase):
         temp = [test_input[k] for k in sorted(test_input)]
         total_array = np.concatenate(temp)
 
-        result = SVM_Classical(training_input, test_input, total_array).run()
-
-        self.assertEqual(result['testing_accuracy'], 1.0)
-        self.assertEqual(result['predicted_classes'],
-                         ['A', 'A', 'A', 'A', 'A', 'A', 'A', 'A', 'A', 'A',
-                          'B', 'B', 'B', 'B', 'B', 'B', 'B', 'B', 'B', 'B'])
+        try:
+            result = SVM_Classical(training_input, test_input, total_array).run()
+            self.assertEqual(result['testing_accuracy'], 1.0)
+            self.assertEqual(result['predicted_classes'],
+                             ['A', 'A', 'A', 'A', 'A', 'A', 'A', 'A', 'A', 'A',
+                              'B', 'B', 'B', 'B', 'B', 'B', 'B', 'B', 'B', 'B'])
+        except NameError as ex:
+            self.skipTest(str(ex))
 
     def test_classical_multiclass_one_against_all(self):
         """ classical multiclass one against all test """